+++ src/util.rs
@@ -64,7 +64,7 @@ pub const MAX_BANNER: ImageSizeLimit = ImageSizeLimit {
 pub const MAX_CSS_LEN: usize = 1024 * 50;
 
 pub const AUTHTOKEN_COOKIE: &str = "token";
-pub const AUTHTOKEN_TTL: usize = 24 * 60 * 60 * 7;
+pub const AUTHTOKEN_TTL: u64 = 24 * 60 * 60 * 7;
 pub const AUTHTOKEN_TTL_I64: i64 = 24 * 60 * 60 * 7;
 
 #[derive(Debug, Copy, Clone)]
